dracut-core: demote binutils to suggests
The package `binutils` provides the binaries `strip`, `objcopy`,
`strings`.

Stripping the binaries in the initrd is an optional feature and not
needed for the files shipped by our Debian packages since those are
already stripped.

`objcopy` and `strings` are only needed when using UEFI executables.

So demote binutils from `Recommends` to `Suggests`.

--CL--

      - demote binutils to suggests. its only needed when using UEFI executables