Revert "Drop dracut-config-generic"
This reverts commit 11b881b66ea4377069ace16a0761981d8e6d490a.

In Ubuntu, initramfs-tools builds portable initrds by default. We want
that user can replace hardware (e.g. the graphics card, keyboard) and
still be able to boot. Taking a disk from one machine, plug it into a
different machine, and boot should work out of the box.

That was the reason Ubuntu sets `hostonly=false`. Pull request #746 and
issue #674 revealed that I misunderstood the `no-hostonly` mode and I
had a lack of knowledge. The man page says in "Boot parameters":

> An initramfs generated without the "hostonly" mode, does not contain
> any system configuration files (except for some special exceptions),
> so the configuration has to be done on the kernel command line.

But as you can see in #746 we want to honor the system configuration
files in Ubuntu.

So restore the dracut-config-generic package.

See also https://github.com/dracut-ng/dracut-ng/issues/748