dracut-test: depend on mount for mount/umount commands
The Dracut test modules uses several binaries. Some of the binaries are
provided by the essential packages coreutils, grep, and util-linux.

The dracut-core package depends on e2fsprogs for the `mkfs.ext4` command
and on `systemd-sysv | initscripts | sysvinit-core` for the `poweroff`
command.

Let the dracut-test package depend on mount for the mount and umount
commands.