fix(systemd-sysusers): increase ordering from 68 to 78
This module should be orders afer all modules that depends on it

This is to make sure that all inst_sysusers calls are in place
before systemd-sysusers is called.

This rule is already documented in the module-setup.sh file
but was not followed for all modules.

Fixes: https://github.com/dracut-ng/dracut-ng/issues/1610
Origin: upstream, https://github.com/dracut-ng/dracut-ng/pull/1613

--CL--

      - fix(systemd-sysusers): increase ordering from 68 to 78