fix(resume): avoid matching extra resume= substrings in cmdline
`grep ... 'resume='` matches not only 'resume' parameter, but any alike
substring. It could be other parameter ending in 'resume' or even parameter
value. The presense of those could fail the module setup.
Fix it by checking for a space or the beginning of a line.

Fixes: https://github.com/dracut-ng/dracut-ng/issues/1595
Signed-off-by: Alexander Zubkov <green@qrator.net>
Origin: upstream, https://github.com/dracut-ng/dracut-ng/pull/1615

--CL--

      - fix(resume): avoid matching extra resume= substrings in cmdline