test: use UEFI for QEMU when available and support UEFI on ARM
Move the `ovmf_code` function and related code to `run-qemu` and always
use UEFI when available.

The OVMF firmware files are builds of EDK II for 64-bit x86 virtual
machines. They do not work on other architectures.

So search for AAVMF files on ARM. Remove the `cfi.pflash01` driver
setting because it does not work on Ubuntu arm64 and `readonly=on` is
already set for the firmware file.

Forwarded: https://github.com/dracut-ng/dracut-ng/pull/1872