autopkgtest: skip too slow 41-full-systemd on arm64 and armhf
The dracut autopkgtests are very slow on arm64 and armhf. They execute
much faster on Debian and on the upstream GitHub CI (upstream tests
arm64, but not armhf).

Taking the simple and fast 10-basic test as reference:

* LP Ubuntu amd64: 13 s
* LP Ubuntu arm64: 304 s
* LP Ubuntu armhf: 282 s
* LP Ubuntu ppc64el: 94 s
* LP Ubuntu s390x: 53 s
* debci Debian amd64: 37 s
* debci Debian arm64: 90 s
* debci Debian ppc64el: 122 s
* debci Debian s390x: 104 s
* Upstream Ubuntu amd64: 22 s
* Upstream Ubuntu arm64: 43 s
* Upstream Debian amd64: 21 s
* Upstream Debian arm64: 34 s

These numbers translate to the total execution time which is one hour on amd64.

41-full-systemd runs into a time out due to this slow execution:

```
[ 164.819978] systemd[1]: Timed out waiting for device dev-disk-by\x2dlabel-dracutusr.device - /dev/disk/by-label/dracutusr.
```

Bug-Ubuntu: https://launchpad.net/bugs/2133401

--CL--

  * autopkgtest: skip too slow 41-full-systemd on arm64 and armhf
    (see LP 2133401)